refactored so TodoContainer connects to Redux store. passing all tests
