+++ src/containers/TodoContainer.js
@@ -1,6 +1,8 @@
 import React, { PropTypes } from 'react';
+import { connect } from 'react-redux';
 import TodoHeader from './../components/TodoHeader';
 import TodoList from './../components/TodoList';
+import * as actions from './../actions/todoActions';
 
 const TodoContainer = (props) => (
   <div>
@@ -17,6 +19,30 @@ const TodoContainer = (props) => (
   </div>
 );
 
+const mapStateToProps = (state) => {
+  return {
+    tasks: state.todo.tasks,
+    taskName: state.todo.taskName,
+  };
+};
+
+const mapDispatchToProps = (dispatch) => {
+  return {
+    changeTaskName: (taskName) => {
+      dispatch(actions.updateTaskActionCreator(taskName));
+    },
+    addTask: (task) => {
+      dispatch(actions.addTaskActionCreator(task));
+    },
+    toggleDone: (index) => {
+      dispatch(actions.markTaskActionCreator(index));
+    },
+    deleteTask: (index) => {
+      dispatch(actions.deleteTaskActionCreator(index));
+    },
+  };
+};
+
 TodoContainer.propTypes = {
   tasks: PropTypes.arrayOf(PropTypes.object),
   taskName: PropTypes.string,
@@ -26,4 +52,4 @@ TodoContainer.propTypes = {
   deleteTask: PropTypes.func,
 };
 
-export default TodoContainer;
+export default connect(mapStateToProps, mapDispatchToProps)(TodoContainer);
